+++ src/components/multipleSemesterTable.tsx
@@ -8,7 +8,6 @@ import { Plan } from "../interfaces/plan";
 export function MultipleSemesterTable({
     plan,
     semester,
-    updateEditedCourse,
     updateDeletedCourse,
     plans,
     setPlans
@@ -17,42 +16,12 @@ export function MultipleSemesterTable({
     semester: Semester;
     plans: Plan[];
     setPlans: (p: Plan[]) => void;
-    updateEditedCourse: (
-        planId: string,
-        semId: string,
-        courseCode: string,
-        newCode: string,
-        newName: string,
-        newCredits: string
-    ) => void;
     updateDeletedCourse: (
         planId: string,
         semId: string,
         courseCode: string
     ) => void;
 }): JSX.Element {
-    //this function will return all the courses; organized by semester in a single plan
-    //im changing this to be a single semester passed in.
-    //this is becuase I have semesterView calling this, which is already mapping an array of semesters
-    //HYPOTHETICALLY, it should do the same thing
-    //list of semesters
-    //const SEMESTER = plan.semesters.map((sem: Semester) => ({ ...sem }));
-    /*
-    return (
-        <div className="container">
-            {SEMESTER.map((sem: Semester) => (
-                <div key={sem.id}>
-
-                    <h4 className="semester">
-                        {sem.season} {sem.year}
-                    </h4>
-                    <h4 className="semester">{sem.id}</h4>
-                    <CourseList semester={sem}></CourseList>
-                </div>
-            ))}
-        </div>
-    );
-    */
     return (
         <div>
             <h4 className="semester">
@@ -63,7 +32,6 @@ export function MultipleSemesterTable({
                 setPlans={setPlans}
                 plan={plan}
                 semester={semester}
-                updateEditedCourse={updateEditedCourse}
                 updateDeletedCourse={updateDeletedCourse}
             ></CourseList>
         </div>
